架构：🚧 改为事件驱动设计模式

+++ app/Services/UserService.php
@@ -2,7 +2,6 @@
 
 namespace App\Services;
 
-use App\Models\DeviceTrack;
 use App\Models\User;
 
 class UserService
@@ -12,16 +11,11 @@ class UserService
      *
      * @param $user_id
      */
-    public static function deleteUser($user_id)
+    public static function userDelete($user_id)
     {
         $user = User::where('id', $user_id)->first();
-        $device_tracks = DeviceTrack::where('user_id', $user_id)
-            ->get();
-
-        foreach ($device_tracks as $device_track) {
-            $device_track->delete();
+        if (!empty($user)) {
+            $user->delete();
         }
-
-        $user->delete();
     }
 }
